Expand invalid block ID testing; move to separate file
